requirements.txt
# Base de l'application
Flask==3.0.2
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
email-validator==2.1.1
mysql-connector-python==8.3.0
python-dotenv==1.0.0

# Sécurité et Nettoyage (Ajoutés pour app.py)
bleach==6.1.0
flask-limiter==3.5.0

# Outils pour la Pipeline CI/CD
flake8==7.0.0
bandit==1.7.7
pip-audit==2.7.2

Flask-Limiter
Flask-Talisman
pyOpenSSL